requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-aiplatform
google-cloud-storage
google-cloud-bigquery